requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
anthropic==0.25.0
openai==1.28.0
hdbscan==0.8.38.post1
scikit-learn==1.4.2
numpy==1.26.4
spacy==3.7.5
python-dotenv==1.0.1
cryptography==42.0.7
reportlab==4.1.0
resend==2.0.0
stripe==9.9.0
apscheduler==3.10.4
rich==13.7.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
